ui/requirements.txt
streamlit>=1.29.0
requests>=2.31.0
pandas>=2.0.0
plotly>=5.17.0
numpy>=1.24.0
python-dotenv>=1.0.0 